Add files via upload
This project is a part of Internshala Ethical Hacking Training. We were provided with an E-commerce website in which we had to find the vulnerabilities that we were taught and are given in the problem statement as well . After submitting our project , we were provided with the solution which I have uploaded.